Move WebAuthnYubicoCredentialsRepository to service
